Write event values for go trials following stop trials
Modify how events are read and written to output a value of one
when there is a go trial following a no-go trial, and output a value
of zero for any other go trial, to try to model the learning on
go trials immediately following each no-go trial. The logic behind this
type of model is that participants are more vigilant after a no-go trial,
and will be slower to react to go trials.